requirements.txt
logging~=0.4.9.6
rsa~=4.9
requests~=2.31.0
python-dotenv~=1.0.1